cleaning up payments package
