+++ packages/payments/src/lib/stripe.ts
@@ -1,0 +1,8 @@
+import Stripe from "stripe";
+
+export const stripe = new Stripe(process.env.STRIPE_SECRET_KEY as string, {
+  // https://github.com/stripe/stripe-node#configuration
+  apiVersion: "2023-10-16",
+});
+
+export { Stripe } from "stripe";
