+++ packages/payments/src/server.ts
@@ -1,168 +1,4 @@
-import {
-  calculateTrialEndUnixTimestamp,
-  createOrRetrieveCustomer,
-} from "./helpers";
-import { getSession } from "@startupkit/auth/server";
-import { getErrorRedirect, getURL } from "./url";
-import { type Price, PricingType } from "@prisma/client";
-
-import Stripe from "stripe";
-
-export const stripe = new Stripe(process.env.STRIPE_SECRET_KEY as string, {
-  // https://github.com/stripe/stripe-node#configuration
-  apiVersion: "2023-10-16",
-});
-
-type CheckoutResponse = {
-  errorRedirect?: string;
-  sessionId?: string;
-};
-
-export async function checkoutWithStripe(
-  price: Price,
-  redirectPath: string = "/account",
-): Promise<CheckoutResponse> {
-  try {
-    const { user } = await getSession();
-
-    if (!user) {
-      throw new Error("Could not get user session.");
-    }
-
-    // Retrieve or create the customer in Stripe
-    let customer: string;
-    try {
-      customer = await createOrRetrieveCustomer({
-        uuid: user?.id || "",
-        email: user?.email || "",
-      });
-    } catch (err) {
-      console.error(err);
-      throw new Error("Unable to access customer record.");
-    }
-
-    let params: Stripe.Checkout.SessionCreateParams = {
-      allow_promotion_codes: true,
-      billing_address_collection: "required",
-      customer,
-      customer_update: {
-        address: "auto",
-      },
-      line_items: [
-        {
-          price: price.id,
-          quantity: 1,
-        },
-      ],
-      cancel_url: getURL(),
-      success_url: getURL(redirectPath),
-    };
-
-    console.log(
-      "Trial end:",
-      calculateTrialEndUnixTimestamp(price.trialPeriodDays),
-    );
-    if (price.type === PricingType.RECURRING) {
-      params = {
-        ...params,
-        mode: "subscription",
-        subscription_data: {
-          trial_end: calculateTrialEndUnixTimestamp(price.trialPeriodDays),
-        },
-      };
-    } else if (price.type === PricingType.ONE_TIME) {
-      params = {
-        ...params,
-        mode: "payment",
-      };
-    }
-
-    // Create a checkout session in Stripe
-    let session;
-    try {
-      session = await stripe.checkout.sessions.create(params);
-    } catch (err) {
-      console.error(err);
-      throw new Error("Unable to create checkout session.");
-    }
-
-    // Instead of returning a Response, just return the data or error.
-    if (session) {
-      return { sessionId: session.id };
-    } else {
-      throw new Error("Unable to create checkout session.");
-    }
-  } catch (error) {
-    if (error instanceof Error) {
-      return {
-        errorRedirect: getErrorRedirect(
-          redirectPath,
-          error.message,
-          "Please try again later or contact a system administrator.",
-        ),
-      };
-    } else {
-      return {
-        errorRedirect: getErrorRedirect(
-          redirectPath,
-          "An unknown error occurred.",
-          "Please try again later or contact a system administrator.",
-        ),
-      };
-    }
-  }
-}
-
-export async function createStripePortal(currentPath: string) {
-  try {
-    const { user } = await getSession();
-
-    if (!user) {
-      throw new Error("Could not get user session.");
-    }
-
-    let customer;
-    try {
-      customer = await createOrRetrieveCustomer({
-        uuid: user.id || "",
-        email: user.email || "",
-      });
-    } catch (err) {
-      console.error(err);
-      throw new Error("Unable to access customer record.");
-    }
-
-    if (!customer) {
-      throw new Error("Could not get customer.");
-    }
-
-    try {
-      const { url } = await stripe.billingPortal.sessions.create({
-        customer,
-        return_url: getURL("/account"),
-      });
-      if (!url) {
-        throw new Error("Could not create billing portal");
-      }
-      return url;
-    } catch (err) {
-      console.error(err);
-      throw new Error("Could not create billing portal");
-    }
-  } catch (error) {
-    if (error instanceof Error) {
-      console.error(error);
-      return getErrorRedirect(
-        currentPath,
-        error.message,
-        "Please try again later or contact a system administrator.",
-      );
-    } else {
-      return getErrorRedirect(
-        currentPath,
-        "An unknown error occurred.",
-        "Please try again later or contact a system administrator.",
-      );
-    }
-  }
-}
+export * from "./server/checkoutWithStripe";
+export * from "./server/getPortalLink";
+export * from "./server/getProducts";
+export * from "./server/getSubscription";
