+++ packages/payments/src/server/checkoutWithStripe.ts
@@ -1,0 +1,107 @@
+import Stripe from "stripe";
+import {
+  calculateTrialEndUnixTimestamp,
+  createOrRetrieveCustomer,
+} from "../helpers";
+import { getErrorRedirect, getURL } from "../lib/url";
+import { type Price, PricingType } from "@prisma/client";
+import { stripe } from "../lib/stripe";
+
+type CheckoutResponse = {
+  errorRedirect?: string;
+  sessionId?: string;
+};
+
+export async function checkoutWithStripe(
+  user: { id: string; email: string },
+  price: Price,
+  redirectPath: string = "/account",
+): Promise<CheckoutResponse> {
+  try {
+    if (!user) {
+      throw new Error("Could not get user session.");
+    }
+
+    // Retrieve or create the customer in Stripe
+    let customer: string;
+    try {
+      customer = await createOrRetrieveCustomer({
+        uuid: user?.id || "",
+        email: user?.email || "",
+      });
+    } catch (err) {
+      console.error(err);
+      throw new Error("Unable to access customer record.");
+    }
+
+    let params: Stripe.Checkout.SessionCreateParams = {
+      allow_promotion_codes: true,
+      billing_address_collection: "required",
+      customer,
+      customer_update: {
+        address: "auto",
+      },
+      line_items: [
+        {
+          price: price.id,
+          quantity: 1,
+        },
+      ],
+      cancel_url: getURL(),
+      success_url: getURL(redirectPath),
+    };
+
+    console.log(
+      "Trial end:",
+      calculateTrialEndUnixTimestamp(price.trialPeriodDays),
+    );
+    if (price.type === PricingType.RECURRING) {
+      params = {
+        ...params,
+        mode: "subscription",
+        subscription_data: {
+          trial_end: calculateTrialEndUnixTimestamp(price.trialPeriodDays),
+        },
+      };
+    } else if (price.type === PricingType.ONE_TIME) {
+      params = {
+        ...params,
+        mode: "payment",
+      };
+    }
+
+    // Create a checkout session in Stripe
+    let session;
+    try {
+      session = await stripe.checkout.sessions.create(params);
+    } catch (err) {
+      console.error(err);
+      throw new Error("Unable to create checkout session.");
+    }
+
+    // Instead of returning a Response, just return the data or error.
+    if (session) {
+      return { sessionId: session.id };
+    } else {
+      throw new Error("Unable to create checkout session.");
+    }
+  } catch (error) {
+    if (error instanceof Error) {
+      return {
+        errorRedirect: getErrorRedirect(
+          redirectPath,
+          error.message,
+          "Please try again later or contact a system administrator.",
+        ),
+      };
+    } else {
+      return {
+        errorRedirect: getErrorRedirect(
+          redirectPath,
+          "An unknown error occurred.",
+          "Please try again later or contact a system administrator.",
+        ),
+      };
+    }
+  }
+}
