+++ packages/payments/src/server/getProducts.ts
@@ -1,0 +1,18 @@
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
+
+export async function getProducts() {
+  return prisma.product.findMany({
+    where: {
+      active: true,
+    },
+    include: {
+      prices: {
+        where: {
+          active: true,
+        },
+      },
+    },
+  });
+}
