+++ packages/payments/src/server/getSubscription.ts
@@ -1,0 +1,23 @@
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
+
+export async function getSubscription(userId: string) {
+  if (!userId) return null;
+
+  return prisma.subscription.findFirst({
+    where: {
+      userId,
+      status: {
+        in: ["trialing", "active"],
+      },
+    },
+    include: {
+      price: {
+        include: {
+          product: true,
+        },
+      },
+    },
+  });
+}
